propbot/requirements.txt
# Web scraping
scrapingbee==1.1.1
requests>=2.26.0
beautifulsoup4>=4.10.0
lxml>=4.6.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2

# Web interface
flask>=2.0.0
flask-cors>=3.0.0
jinja2==3.1.2

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
python-dotenv>=0.19.0
tqdm==4.66.1
schedule==1.2.1

# PropBot Dependencies
jsonschema>=4.17.0
pathlib>=1.0.1
typing-extensions>=4.2.0
python-dateutil>=2.8.2
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0

